hw5/requirements.txt
# ctc-decoder==1.0.1 install this package by https://github.com/githubharald/CTCDecoder
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.10.0.84
pandas==2.0.3
Pillow==7.0.0
seaborn==0.13.2
torchvision==0.19.1  # install torch by torchvision
trdg==1.6.0